+++ sql/emaj--2.2.1--next_version.sql
@@ -1,0 +1,3 @@
+--
+-- E-Maj: migration from 2.2.1 to <NEXT_VERSION>
+--